Move wrapper types to a submodule.
