+++ src/lib.rs
@@ -10,41 +10,15 @@
 
 #[cfg(target_arch = "aarch64")]
 mod aarch64_mmio;
+pub mod fields;
 mod physical;
 #[cfg(not(target_arch = "aarch64"))]
 mod volatile_mmio;
 
+use crate::fields::{ReadOnly, ReadPure, ReadPureWrite, ReadWrite, WriteOnly};
 use core::{array, fmt::Debug, marker::PhantomData, ops::Deref, ptr::NonNull};
 pub use physical::PhysicalInstance;
-use zerocopy::{FromBytes, Immutable, IntoBytes, KnownLayout};
-
-/// Wrapper for a field which may safely be read but not written. Reading may cause side-effects,
-/// changing the state of the device in some way.
-#[derive(Clone, Debug, Eq, FromBytes, Immutable, IntoBytes, KnownLayout, PartialEq)]
-#[repr(transparent)]
-pub struct ReadOnly<T>(pub T);
-
-/// Wrapper for a field which may safely be read with no side-effects but not written.
-#[derive(Clone, Debug, Eq, FromBytes, Immutable, IntoBytes, KnownLayout, PartialEq)]
-#[repr(transparent)]
-pub struct ReadPure<T>(pub T);
-
-/// Wrapper for a field which may safely be written but not read.
-#[derive(Clone, Debug, Eq, FromBytes, Immutable, IntoBytes, KnownLayout, PartialEq)]
-#[repr(transparent)]
-pub struct WriteOnly<T>(pub T);
-
-/// Wrapper for a field which may safely be written and read. Reading may cause side-effects,
-/// changing the state of the device in some way.
-#[derive(Clone, Debug, Eq, FromBytes, Immutable, IntoBytes, KnownLayout, PartialEq)]
-#[repr(transparent)]
-pub struct ReadWrite<T>(pub T);
-
-/// Wrapper for a field which may safely be written (with side-effects) and read with no
-/// side-effects.
-#[derive(Clone, Debug, Eq, FromBytes, Immutable, IntoBytes, PartialEq)]
-#[repr(transparent)]
-pub struct ReadPureWrite<T>(pub T);
+use zerocopy::{FromBytes, Immutable, IntoBytes};
 
 /// A unique owned pointer to the registers of some MMIO device.
 ///
